dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_DOMImplementation-hasFeature.html

<!doctype html>
<title>DOMImplementation.hasFeature(feature, version)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var tests = [
        ["Core", "1.0", false],
        ["Core", "2.0", true],
        ["Core", "3.0", false],
        ["Core", "100.0", false],
        ["XML", "1.0", true],
        ["XML", "2.0", true],
        ["XML", "3.0", false],
        ["XML", "100.0", false],
        ["Core", "1", false],
        ["Core", "2", false],
        ["Core", "3", false],
        ["Core", "100", false],
        ["XML", "1", false],
        ["XML", "2", false],
        ["XML", "3", false],
        ["XML", "100", false],
        ["Core", "1.1", false],
        ["Core", "2.1", false],
        ["Core", "3.1", false],
        ["Core", "100.1", false],
        ["XML", "1.1", false],
        ["XML", "2.1", false],
        ["XML", "3.1", false],
        ["XML", "100.1", false],
        ["Core", "", true],
        ["XML", "", true],
        ["core", "", true],
        ["xml", "", true],
        ["CoRe", "", true],
        ["XmL", "", true],
        [" Core", "", false],
        [" XML", "", false],
        ["Core ", "", false],
        ["XML ", "", false],
        ["Co re", "", false],
        ["XM L", "", false],
        ["aCore", "", false],
        ["aXML", "", false],
        ["Corea", "", false],
        ["XMLa", "", false],
        ["Coare", "", false],
        ["XMaL", "", false],
        ["Core", " ", false],
        ["XML", " ", false],
        ["Core", " 1.0", false],
        ["Core", " 2.0", false],
        ["Core", " 3.0", false],
        ["Core", " 100.0", false],
        ["XML", " 1.0", false],
        ["XML", " 2.0", false],
        ["XML", " 3.0", false],
        ["XML", " 100.0", false],
        ["Core", "1.0 ", false],
        ["Core", "2.0 ", false],
        ["Core", "3.0 ", false],
        ["Core", "100.0 ", false],
        ["XML", "1.0 ", false],
        ["XML", "2.0 ", false],
        ["XML", "3.0 ", false],
        ["XML", "100.0 ", false],
        ["Core", "1. 0", false],
        ["Core", "2. 0", false],
        ["Core", "3. 0", false],
        ["Core", "100. 0", false],
        ["XML", "1. 0", false],
        ["XML", "2. 0", false],
        ["XML", "3. 0", false],
        ["XML", "100. 0", false],
        ["Core", "a1.0", false],
        ["Core", "a2.0", false],
        ["Core", "a3.0", false],
        ["Core", "a100.0", false],
        ["XML", "a1.0", false],
        ["XML", "a2.0", false],
        ["XML", "a3.0", false],
        ["XML", "a100.0", false],
        ["Core", "1.0a", false],
        ["Core", "2.0a", false],
        ["Core", "3.0a", false],
        ["Core", "100.0a", false],
        ["XML", "1.0a", false],
        ["XML", "2.0a", false],
        ["XML", "3.0a", false],
        ["XML", "100.0a", false],
        ["Core", "1.a0", false],
        ["Core", "2.a0", false],
        ["Core", "3.a0", false],
        ["Core", "100.a0", false],
        ["XML", "1.a0", false],
        ["XML", "2.a0", false],
        ["XML", "3.a0", false],
        ["XML", "100.a0", false],
        ["Core", 1, false],
        ["Core", 2, false],
        ["Core", 3, false],
        ["Core", 100, false],
        ["XML", 1, false],
        ["XML", 2, false],
        ["XML", 3, false],
        ["XML", 100, false],
        ["Core", null, true],
        ["XML", null, true],
        ["core", null, true],
        ["xml", null, true],
        ["CoRe", null, true],
        ["XmL", null, true],
        [" Core", null, false],
        [" XML", null, false],
        ["Core ", null, false],
        ["XML ", null, false],
        ["Co re", null, false],
        ["XM L", null, false],
        ["aCore", null, false],
        ["aXML", null, false],
        ["Corea", null, false],
        ["XMLa", null, false],
        ["Coare", null, false],
        ["XMaL", null, false],
        ["Core", undefined, false],
        ["XML", undefined, false],
        ["This is filler text.", "", false],
        [null, "", false],
        [undefined, "", false],
      ]
  for (var i in tests) {
    var test = tests[i]
    assert_equals(document.implementation.hasFeature(test[0], test[1]), test[2], test[0] + " " + test[1])
  }
})
</script>
